Web/sql_based_cookie_injection/flask/requirements.txt
click==8.0.1
colorama==0.4.4
dominate==2.6.0
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.15.1
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql==0.0.3
mysqlclient==2.0.3
typing-extensions==3.10.0.2
visitor==0.1.3
waitress==2.0.0
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.5.0
